Move Min/Max instances to companion objects
